inseridas as partes de Controllers,Conexao com o Entity Framework, insercao de DDD, Validadores.

+++ ApiContatos/Infrastructure/Repositories/ContactRepository/IContactRepository.cs
@@ -6,13 +6,13 @@ namespace ApiContatos.Infrastructure.Repositories.ContactRepository
     {
         Task<(IEnumerable<Contact> data, int totalCount)> GetAll(int page, int pageSize);
 
-        Contact GetById(int id);
+        Contact GetById(long id);
 
         void Create(Contact entity);
 
         void Update(Contact entity);
 
-        void Delete(int id);
+        void Delete(long id);
